Refactor settings management by restructuring resources
- Moved SettingResource to a new namespace: Firefly\FilamentBlog\Resources\Settings.
- Introduced new schemas and pages for managing settings, including CreateSetting, EditSetting, and ListSettings.
- Removed the deprecated SettingResource class and its associated methods to enhance clarity and maintainability.
- Updated Blog.php and other related files to reflect the new SettingsResource path.
