--- src/Resources/SettingResource.php
@@ -1,94 +1,0 @@
-<?php
-
-namespace Firefly\FilamentBlog\Resources;
-
-use Filament\Schemas\Schema;
-use Filament\Tables\Columns\TextColumn;
-use Filament\Tables\Columns\ImageColumn;
-use Filament\Actions\EditAction;
-use Filament\Actions\ViewAction;
-use Filament\Actions\BulkActionGroup;
-use Filament\Actions\DeleteBulkAction;
-use Firefly\FilamentBlog\Resources\SettingResource\Pages\ListSettings;
-use Firefly\FilamentBlog\Resources\SettingResource\Pages\CreateSetting;
-use Firefly\FilamentBlog\Resources\SettingResource\Pages\EditSetting;
-use Filament\Resources\Resource;
-use Filament\Tables;
-use Filament\Tables\Table;
-use Firefly\FilamentBlog\Models\Setting;
-
-class SettingResource extends Resource
-{
-    protected static ?string $model = Setting::class;
-
-    protected static string | \BackedEnum | null $navigationIcon = 'heroicon-o-cog-6-tooth';
-
-    protected static string | \UnitEnum | null $navigationGroup = 'Blog';
-
-    protected static ?int $navigationSort = 8;
-
-    public static function form(Schema $schema): Schema
-    {
-        return $schema
-            ->components(Setting::getForm());
-    }
-
-    public static function canCreate(): bool
-    {
-        return Setting::count() === 0;
-    }
-
-    public static function table(Table $table): Table
-    {
-        return $table
-            ->columns([
-                TextColumn::make('title')
-                    ->limit(25)
-                    ->searchable(),
-                TextColumn::make('description')
-                    ->limit(30)
-                    ->searchable(),
-
-                ImageColumn::make('logo'),
-
-                TextColumn::make('organization_name'),
-
-                TextColumn::make('created_at')
-                    ->dateTime()
-                    ->sortable()
-                    ->toggleable(isToggledHiddenByDefault: true),
-                TextColumn::make('updated_at')
-                    ->dateTime()
-                    ->sortable()
-                    ->toggleable(isToggledHiddenByDefault: true),
-            ])
-            ->filters([
-                //
-            ])
-            ->recordActions([
-                EditAction::make(),
-                ViewAction::make(),
-            ])
-            ->toolbarActions([
-                BulkActionGroup::make([
-                    DeleteBulkAction::make(),
-                ]),
-            ]);
-    }
-
-    public static function getRelations(): array
-    {
-        return [
-            //
-        ];
-    }
-
-    public static function getPages(): array
-    {
-        return [
-            'index' => ListSettings::route('/'),
-            'create' => CreateSetting::route('/create'),
-            'edit' => EditSetting::route('/{record}/edit'),
-        ];
-    }
-}
