+++ src/Resources/Settings/Pages/CreateSetting.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Firefly\FilamentBlog\Resources\Settings\Pages;
+
+use Filament\Resources\Pages\CreateRecord;
+use Firefly\FilamentBlog\Resources\Settings\SettingResource;
+
+class CreateSetting extends CreateRecord
+{
+    protected static string $resource = SettingResource::class;
+}
